[Truffle] Make openssl stubs opt-in
