+++ lib/ruby/truffle/jruby+truffle/gem_ci/activesupport.rb
@@ -9,5 +9,5 @@ use_only_https_git_paths!
 
 has_to_succeed setup
 
-result run(%w[--require-pattern test/**/*_test.rb -r exclude_tests -- -I test -e nil], raise: false)
+result run(%w[--require-pattern test/**/**/*_test.rb -r exclude_tests -- -I test -e nil], raise: false)
