+++ test/truffle/integration/rails/.jruby+truffle.yaml
@@ -48,4 +48,5 @@
 
 :run:
   :require:
+    - openssl-stubs
     - stubs
